% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{AAFE}
\alias{AAFE}
\title{Absolute Average-Fold Error}
\usage{
AAFE(predictions, observations, times = NULL)
}
\arguments{
\item{predictions}{list with the predicted values}

\item{observations}{list with the observed values}
}
\description{
Estimates the absolute average-fold error.
}
