% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{AUC}
\alias{AUC}
\title{Area Under the Curb - AUC}
\usage{
AUC(x, y)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
This function estimate the area under the curve of a given dataset
}
