% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_estimation.R
\name{ci_estimation}
\alias{ci_estimation}
\title{Confidence Intervals}
\usage{
ci_estimation(
  pl_results,
  alpha = 0.95,
  df = 1,
  theta_hat,
  global_optimum,
  lb,
  ub
)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
This code has been developed to be used with the profile_likelihood and identifiability_analysis
functions, in order to estimate the confidence intervals of the parameters, based on their
likelihood profiles. The method is an implementation of the methodology described in the
publication of Raue et al., 2009 entitled as "Structural and practical
identifiability analysis of partially observed dynamical models by exploiting
the profile likelihood" (doi: https://doi.org/10.1093/bioinformatics/btp358).
}
