% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifiability_analysis.R
\name{Identifiability_analysis}
\alias{Identifiability_analysis}
\title{Identifiability Analysis}
\usage{
Identifiability_analysis(
  obj_f,
  thetas,
  thetas_names,
  data_df,
  errors_df,
  lb,
  ub,
  N_samples = 50,
  alpha = 0.95,
  df = 1,
  q = 0.5,
  global_optimum,
  min_step_coef = 0.001,
  max_step_coef = 0.2,
  N_cores,
  constant_params = NULL,
  exported_to_cluster = NULL,
  break_at_bounds = FALSE,
  opts = list(algorithm = "NLOPT_LN_NELDERMEAD", xtol_rel = 1e-06, ftol_rel = 1e-06,
    ftol_abs = 0, xtol_abs = 0, maxeval = 300, print_level = 0),
  opts_theta_step = list(algorithm = "NLOPT_LN_SBPLX", xtol_rel_step = 1e-05,
    ftol_rel_step = 1e-05, ftol_abs_step = 0, xtol_abs_step = 0, maxeval_step = 50,
    print_level_step = 0),
  create_txt = TRUE
)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
This is a function to implement the framework for identifiability analysis
as described in the publication of Raue et al., 2009 entitled as "Structural and practical
identifiability analysis of partially observed dynamical models by exploiting
the profile likelihood" (doi: https://doi.org/10.1093/bioinformatics/btp358).
}
