% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(observed, predicted)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
Estimates the mean absolute percentage error.
}
