% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{mse_custom}
\alias{mse_custom}
\title{Mean Squared Error}
\usage{
mse_custom(observed, predicted)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
Estimates the Mean Squared Error
}
