% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{PBKOF}
\alias{PBKOF}
\title{PBK Objective Function}
\usage{
PBKOF(observed, predicted, comp.names = NULL)
}
\arguments{
\item{observed}{list with the observed values}

\item{predicted}{list with the predicted values}
}
\description{
A custom metric for PBK models.
}
