% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(observed, predicted)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
Estimates the root mean squared error.
}
