% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting_func.R
\name{Weighting.func}
\alias{Weighting.func}
\title{Estimating Weights for data points.}
\usage{
Weighting.func(df, a_first = 2, a_last = 2, a_tp = 2)
}
\arguments{
\item{df}{a dataframe whose the 1st column contains the values for the
independent variable (x) and the rest columns contain the data for the dependent variables (y).}

\item{a_first}{the value of the factor to estimate the weight for the first point
of the data}

\item{a_last}{the value of the factor to estimate the weight for the last point
of the data}

\item{a_tp}{value of the factor to estimate the weight for turning points (or the
points that are next to the turning points)}
}
\value{
A list with the weight values for each dependent variable of the dataset.
}
\description{
This custom function is designed to assign weight values to data points within
a dataset. Its primary purpose is to identify turning points or points adjacent
to turning points, as well as the first and last points in the data. These
particular points are considered more significant within the dataset, and as a result,
the function assigns higher weights to these points.
Additionally, the function applies higher weight values to points that are
farther apart from each other. This weighting approach aims to enhance
the model's performance by prioritizing crucial data points and appropriately
accounting for their relative distances.
}
