% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{WSSR}
\alias{WSSR}
\title{Weighted Sum of Squared Residuals}
\usage{
WSSR(observed, predicted, weights, comp.names = NULL)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
Estimates the weighted sum of squared residuals.
}
