% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{AUC}
\alias{AUC}
\title{Area Under the Curb - AUC}
\usage{
AUC(x, y)
}
\arguments{
\item{x}{a vector with the x coordinates}

\item{y}{a vector with the y coordinates}
}
\description{
Estimates the Area Under the Curve.
}
