% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(observed, predicted)
}
\arguments{
\item{observed}{vector with the observed values}

\item{predicted}{vector with the predicted values}
}
\description{
Estimates the mean absolute percentage error.
}
