% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{mse_custom}
\alias{mse_custom}
\title{Mean Squared Error}
\usage{
mse_custom(observed, predicted)
}
\arguments{
\item{observed}{vector with the observed values}

\item{predicted}{vector with the predicted values}
}
\description{
Estimates the Mean Squared Error
}
