% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity_analysis.R
\name{PBPK_sensitivity}
\alias{PBPK_sensitivity}
\title{Local Sensitivity Analysis}
\usage{
PBPK_sensitivity(
  model,
  thetas,
  thetas_names,
  constant_params,
  ranges,
  targets,
  ode_settings,
  heatmap = FALSE
)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
Performs local sensitivity analysis to the parameters of a given model.
}
