% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood.R
\name{profile_likelihood}
\alias{profile_likelihood}
\title{Profile Likelihood}
\usage{
profile_likelihood(
  obj_f,
  i,
  thetas,
  thetas_names,
  constant_params = NULL,
  data_df,
  errors_df,
  lb,
  ub,
  N_samples,
  alpha,
  df,
  q,
  global_optimum,
  min_step_coef,
  max_step_coef,
  break_at_bounds = FALSE,
  opts = list(algorithm = "NLOPT_LN_NELDERMEAD", xtol_rel = 1e-06, ftol_rel = 1e-06,
    ftol_abs = 0, xtol_abs = 0, maxeval = 300, print_level = 1),
  opts_theta_step = list(algorithm = "NLOPT_LN_SBPLX", xtol_rel = 1e-05, ftol_rel =
    1e-05, ftol_abs = 0, xtol_abs = 0, maxeval = 50, print_level = 0),
  create_txt = FALSE
)
}
\arguments{
\item{to}{be completed}
}
\value{
to be completed
}
\description{
This code is an implementation of the Profile Likelihood-based identifiability
analysis, as described
in the publication of Raue et al., 2009 entitled as "Structural and practical
identifiability analysis of partially observed dynamical models by exploiting
the profile likelihood" (doi: https://doi.org/10.1093/bioinformatics/btp358).
}
