% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(observed, predicted)
}
\arguments{
\item{observed}{vector with the observed values}

\item{predicted}{vector with the predicted values}
}
\description{
Estimates the root mean squared error.
}
