% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodness_of_fit_metrics.R
\name{WSSR}
\alias{WSSR}
\title{Weighted Sum of Squared Residuals}
\usage{
WSSR(observed, predicted, weights_values, comp.names = NULL)
}
\arguments{
\item{observed}{list with the observed values}

\item{predicted}{list with the predicted values}

\item{weights_values}{list with the weights}
}
\description{
Estimates the weighted sum of squared residuals.
}
